% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcol_func.R
\name{addcol}
\alias{addcol}
\title{Adding Column Counts Function: Creating a new column counting values in a row across columns}
\usage{
addcol(data, col1, col2)
}
\arguments{
\item{data}{= object you named your data when you read it in}

\item{col1}{= column name where you want your row count to begin}

\item{col2}{= = column name where you want your row count to end}
}
\value{
if new count column is added the tibble/df should be returned with one more column number than your original data frame
}
\description{
Adding Column Counts Function: Creating a new column counting values in a row across columns
}
