% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_func.R
\name{cleandata}
\alias{cleandata}
\title{Data Clean-Up Function removes NA values in the csv data file}
\usage{
cleandata(filepath)
}
\arguments{
\item{filepath}{where your data file is stored in R}
}
\value{
if all NAs are removed then the last line in the console should be \link{1} False, if NAs are not removed it will be \link{1} True
}
\description{
Data Clean-Up Function removes NA values in the csv data file
}
