% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMDSMatrix_func.R
\name{NMDSmatrix}
\alias{NMDSmatrix}
\title{Creating an NMDS Matrix Function: Creating a matrix from your data frame to make an NMDS plot}
\usage{
NMDSmatrix(data, col1, col2, col3)
}
\arguments{
\item{data}{= object you named your data when you read it in}

\item{col1}{= column from the data frame you want to use in your NMDS analysis, such as a column listing the names of your detected species}

\item{col2}{= a different column from the data frame you want to use in your NMDS analysis, such as a count of your detected species}

\item{col3}{= a different column from the others in the data frame you want to use in your NMDS analysis, such as different habitats or other relevent differences}
}
\value{
matrix showing species counts for different habitats or other relevent differences
}
\description{
Creating an NMDS Matrix Function: Creating a matrix from your data frame to make an NMDS plot
}
