% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres_abs_func.R
\name{pres.abs.plot}
\alias{pres.abs.plot}
\title{Presence Absence Function: Visualizing the presence or absence of a species based on other parameters in the data set.}
\usage{
pres.abs.plot(data, x, y, col)
}
\arguments{
\item{data}{= object you named your data when you read it in}

\item{x}{= column from the data frame you want to go on the x axis}

\item{y}{= column from the data frame you want to go on the y axis}

\item{col}{= column from the data frame you want to file the tiles for presence absence data}
}
\value{
ggplot showing presence absence of a species in the data frame
}
\description{
Presence Absence Function: Visualizing the presence or absence of a species based on other parameters in the data set.
}
